removed unused functions
