--- plugin_test.go
@@ -1,7 +1,0 @@
-package gateway
-
-import "testing"
-
-func TestMiddlewareList_queryRequestMiddlewares(t *testing.T) {
-	// build a list of query request Middlewares
-}
